Move PayloadSender to core
