+++ packages/xstate-react/src/types.ts
@@ -46,31 +46,3 @@ export interface ActorRefLike<TEvent extends EventObject, TEmitted = any>
 }
 
 export type MaybeLazy<T> = T | (() => T);
-
-type ExcludeType<A> = { [K in Exclude<keyof A, 'type'>]: A[K] };
-
-type ExtractExtraParameters<A, T> = A extends { type: T }
-  ? ExcludeType<A>
-  : never;
-
-type ExtractSimple<A> = A extends any
-  ? {} extends ExcludeType<A>
-    ? A
-    : never
-  : never;
-
-type NeverIfEmpty<T> = {} extends T ? never : T;
-
-export interface PayloadSender<TEvent extends EventObject> {
-  /**
-   * Send an event object or just the event type, if the event has no other payload
-   */
-  (event: TEvent | ExtractSimple<TEvent>['type']): void;
-  /**
-   * Send an event type and its payload
-   */
-  <K extends TEvent['type']>(
-    eventType: K,
-    payload: NeverIfEmpty<ExtractExtraParameters<TEvent, K>>
-  ): void;
-}
